Almost initial commit
Migrating to GitHub.

+++ HS/Deck.cpp
@@ -20,7 +20,7 @@ Card * Deck::drawRandomCard(std::mt19937_64 &generator)
 	if (numCards == 0) throw EmptyDeckException();
 
 	Card * toDraw = nullptr;
-	int8_t num = generator()%numCards;
+	int32_t num = generator()%numCards;
 
 	toDraw = removeCard(num);
